docker-demo/frontend-streamlit/pyproject.toml
[tool.poetry]
name = "frontend-streamlit"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"
packages = [{include = "frontend_streamlit"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.18.1"
watchdog = "^2.2.1"
requests = "^2.28.2"
python-decouple = "^3.7"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
